Ovet
Molempien ovien pitäisi nyt toimia e-nappia painamalla
