+++ Assets/Scripts/UserInterface/InventoryUIController.cs
@@ -80,4 +80,9 @@ public class InventoryUIController : MonoBehaviour, IPointerEnterHandler, IPoint
     {
         _infoView.SetActive(false);
     }
+    public void OnObject(string ObjectTag)
+    {
+        _infoText.text = "Open/close " + ObjectTag + "  with (E)";
+        _infoView.SetActive(true);
+    }
 }
